+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOTransportConfig/MSFT_EXOTransportConfig.psm1
@@ -1,0 +1,692 @@
+function Get-TargetResource
+{
+    [CmdletBinding()]
+    [OutputType([System.Collections.Hashtable])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateSet('Yes')]
+        [System.String]
+        $IsSingleInstance,
+
+        [Parameter()]
+        [System.Boolean]
+        $AddressBookPolicyRoutingEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowLegacyTLSClients,
+
+        [Parameter()]
+        [System.Boolean]
+        $ClearCategories,
+
+        [Parameter()]
+        [System.Boolean]
+        $ConvertDisclaimerWrapperToEml,
+
+        [Parameter()]
+        [System.String]
+        $DSNConversionMode,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.String]
+        $ExternalDsnDefaultLanguage,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.String]
+        $ExternalDsnReportingAuthority,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDsnSendHtml,
+
+        [Parameter()]
+        [System.String]
+        $ExternalPostmasterAddress,
+
+        [Parameter()]
+        [System.String]
+        $HeaderPromotionModeSetting,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.String]
+        $InternalDsnDefaultLanguage,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.String]
+        $InternalDsnReportingAuthority,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDsnSendHtml,
+
+        [Parameter()]
+        [System.Int32]
+        $JournalMessageExpirationDays,
+
+        [Parameter()]
+        [System.String]
+        $JournalingReportNdrTo,
+
+        [Parameter()]
+        [System.String]
+        $MaxRecipientEnvelopeLimit,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormBlockDurationHours,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormDetectionMinimumRecipients,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormDetectionMinimumReplies,
+
+        [Parameter()]
+        [System.Boolean]
+        $ReplyAllStormProtectionEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $Rfc2231EncodingEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $SmtpClientAuthenticationDisabled,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $Credential,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword
+    )
+
+    Write-Verbose -Message "Getting EXOTransportConfig"
+    if ($Global:CurrentModeIsExport)
+    {
+        $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
+            -InboundParameters $PSBoundParameters `
+            -SkipModuleReload $true
+    }
+    else
+    {
+        $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
+            -InboundParameters $PSBoundParameters
+    }
+
+    #Ensure the proper dependencies are installed in the current environment.
+    Confirm-M365DSCDependencies
+
+    #region Telemetry
+    $ResourceName = $MyInvocation.MyCommand.ModuleName -replace "MSFT_", ""
+    $CommandName = $MyInvocation.MyCommand
+    $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
+        -CommandName $CommandName `
+        -Parameters $PSBoundParameters
+    Add-M365DSCTelemetryEvent -Data $data
+    #endregion
+
+    $nullReturn = $PSBoundParameters
+    $nullReturn.Ensure = "Absent"
+    try
+    {
+        $TransportConfigSettings = Get-TransportConfig -ErrorAction Stop
+        if ($null -eq $TransportConfigSettings)
+        {
+            throw 'There was an error retrieving values from the Get function in EXOTransportConfig.'
+        }
+
+        $results = @{
+            IsSingleInstance                        = 'Yes'
+            AddressBookPolicyRoutingEnabled         = $TransportConfigSettings.AddressBookPolicyRoutingEnabled
+            AllowLegacyTLSClients                   = $TransportConfigSettings.AllowLegacyTLSClients
+            ClearCategories                         = $TransportConfigSettings.ClearCategories
+            ConvertDisclaimerWrapperToEml           = $TransportConfigSettings.ConvertDisclaimerWrapperToEml
+            DSNConversionMode                       = $TransportConfigSettings.DSNConversionMode
+            ExternalDelayDsnEnabled                 = $TransportConfigSettings.ExternalDelayDsnEnabled
+            ExternalDsnDefaultLanguage              = $TransportConfigSettings.ExternalDsnDefaultLanguage
+            ExternalDsnLanguageDetectionEnabled     = $TransportConfigSettings.ExternalDsnLanguageDetectionEnabled
+            ExternalDsnReportingAuthority           = $TransportConfigSettings.ExternalDsnReportingAuthority
+            ExternalDsnSendHtml                     = $TransportConfigSettings.ExternalDsnSendHtml
+            ExternalPostmasterAddress               = $TransportConfigSettings.ExternalPostmasterAddress
+            HeaderPromotionModeSetting              = $TransportConfigSettings.HeaderPromotionModeSetting
+            InternalDelayDsnEnabled                 = $TransportConfigSettings.InternalDelayDsnEnabled
+            InternalDsnDefaultLanguage              = $TransportConfigSettings.InternalDsnDefaultLanguage
+            InternalDsnLanguageDetectionEnabled     = $TransportConfigSettings.InternalDsnLanguageDetectionEnabled
+            InternalDsnReportingAuthority           = $TransportConfigSettings.InternalDsnReportingAuthority
+            InternalDsnSendHtml                     = $TransportConfigSettings.InternalDsnSendHtml
+            JournalMessageExpirationDays            = $TransportConfigSettings.JournalMessageExpirationDays
+            JournalingReportNdrTo                   = $TransportConfigSettings.JournalingReportNdrTo
+            MaxRecipientEnvelopeLimit               = $TransportConfigSettings.MaxRecipientEnvelopeLimit
+            ReplyAllStormBlockDurationHours         = $TransportConfigSettings.ReplyAllStormBlockDurationHours
+            ReplyAllStormDetectionMinimumRecipients = $TransportConfigSettings.ReplyAllStormDetectionMinimumRecipients
+            ReplyAllStormDetectionMinimumReplies    = $TransportConfigSettings.ReplyAllStormDetectionMinimumReplies
+            ReplyAllStormProtectionEnabled          = $TransportConfigSettings.ReplyAllStormProtectionEnabled
+            Rfc2231EncodingEnabled                  = $TransportConfigSettings.Rfc2231EncodingEnabled
+            SmtpClientAuthenticationDisabled        = $TransportConfigSettings.SmtpClientAuthenticationDisabled
+            Credential                              = $Credential
+            ApplicationId                           = $ApplicationId
+            CertificateThumbprint                   = $CertificateThumbprint
+            CertificatePath                         = $CertificatePath
+            CertificatePassword                     = $CertificatePassword
+            TenantId                                = $TenantId
+        }
+
+        return $results
+    }
+    catch
+    {
+        try
+        {
+            Write-Verbose -Message $_
+            $tenantIdValue = ""
+            if (-not [System.String]::IsNullOrEmpty($TenantId))
+            {
+                $tenantIdValue = $TenantId
+            }
+            elseif ($null -ne $Credential)
+            {
+                $tenantIdValue = $Credential.UserName.Split('@')[1]
+            }
+            Add-M365DSCEvent -Message $_ -EntryType 'Error' `
+                -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
+                -TenantId $tenantIdValue
+        }
+        catch
+        {
+            Write-Verbose -Message $_
+        }
+        return $nullReturn
+    }
+}
+
+function Set-TargetResource
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateSet('Yes')]
+        [System.String]
+        $IsSingleInstance,
+
+        [Parameter()]
+        [System.Boolean]
+        $AddressBookPolicyRoutingEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowLegacyTLSClients,
+
+        [Parameter()]
+        [System.Boolean]
+        $ClearCategories,
+
+        [Parameter()]
+        [System.Boolean]
+        $ConvertDisclaimerWrapperToEml,
+
+        [Parameter()]
+        [System.String]
+        $DSNConversionMode,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.String]
+        $ExternalDsnDefaultLanguage,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.String]
+        $ExternalDsnReportingAuthority,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDsnSendHtml,
+
+        [Parameter()]
+        [System.String]
+        $ExternalPostmasterAddress,
+
+        [Parameter()]
+        [System.String]
+        $HeaderPromotionModeSetting,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.String]
+        $InternalDsnDefaultLanguage,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.String]
+        $InternalDsnReportingAuthority,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDsnSendHtml,
+
+        [Parameter()]
+        [System.Int32]
+        $JournalMessageExpirationDays,
+
+        [Parameter()]
+        [System.String]
+        $JournalingReportNdrTo,
+
+        [Parameter()]
+        [System.String]
+        $MaxRecipientEnvelopeLimit,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormBlockDurationHours,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormDetectionMinimumRecipients,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormDetectionMinimumReplies,
+
+        [Parameter()]
+        [System.Boolean]
+        $ReplyAllStormProtectionEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $Rfc2231EncodingEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $SmtpClientAuthenticationDisabled,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $Credential,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword
+    )
+    #Ensure the proper dependencies are installed in the current environment.
+    Confirm-M365DSCDependencies
+
+    #region Telemetry
+    $ResourceName = $MyInvocation.MyCommand.ModuleName -replace "MSFT_", ""
+    $CommandName = $MyInvocation.MyCommand
+    $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
+        -CommandName $CommandName `
+        -Parameters $PSBoundParameters
+    Add-M365DSCTelemetryEvent -Data $data
+    #endregion
+
+    Write-Verbose -Message "Setting EXOTransportConfig"
+
+    $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
+        -InboundParameters $PSBoundParameters
+
+
+    Write-Verbose -Message "Setting EXOTransportConfig with values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
+    $SetValues = [System.Collections.Hashtable]($PSBoundParameters)
+    $SetValues.Remove('IsSingleInstance') | Out-Null
+    $SetValues.Remove('Credential') | Out-Null
+    $SetValues.Remove('ApplicationId') | Out-Null
+    $SetValues.Remove('TenantId') | Out-Null
+    $SetValues.Remove('CertificateThumbprint') | Out-Null
+    $SetValues.Remove('CertificatePath') | Out-Null
+    $SetValues.Remove('CertificatePassword') | Out-Null
+
+    Set-TransportConfig @SetValues
+}
+
+function Test-TargetResource
+{
+    [CmdletBinding()]
+    [OutputType([System.Boolean])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateSet('Yes')]
+        [System.String]
+        $IsSingleInstance,
+
+        [Parameter()]
+        [System.Boolean]
+        $AddressBookPolicyRoutingEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowLegacyTLSClients,
+
+        [Parameter()]
+        [System.Boolean]
+        $ClearCategories,
+
+        [Parameter()]
+        [System.Boolean]
+        $ConvertDisclaimerWrapperToEml,
+
+        [Parameter()]
+        [System.String]
+        $DSNConversionMode,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.String]
+        $ExternalDsnDefaultLanguage,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.String]
+        $ExternalDsnReportingAuthority,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExternalDsnSendHtml,
+
+        [Parameter()]
+        [System.String]
+        $ExternalPostmasterAddress,
+
+        [Parameter()]
+        [System.String]
+        $HeaderPromotionModeSetting,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.String]
+        $InternalDsnDefaultLanguage,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.String]
+        $InternalDsnReportingAuthority,
+
+        [Parameter()]
+        [System.Boolean]
+        $InternalDsnSendHtml,
+
+        [Parameter()]
+        [System.Int32]
+        $JournalMessageExpirationDays,
+
+        [Parameter()]
+        [System.String]
+        $JournalingReportNdrTo,
+
+        [Parameter()]
+        [System.String]
+        $MaxRecipientEnvelopeLimit,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormBlockDurationHours,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormDetectionMinimumRecipients,
+
+        [Parameter()]
+        [System.Int32]
+        $ReplyAllStormDetectionMinimumReplies,
+
+        [Parameter()]
+        [System.Boolean]
+        $ReplyAllStormProtectionEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $Rfc2231EncodingEnabled,
+
+        [Parameter()]
+        [System.Boolean]
+        $SmtpClientAuthenticationDisabled,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $Credential,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword
+    )
+    #Ensure the proper dependencies are installed in the current environment.
+    Confirm-M365DSCDependencies
+
+    #region Telemetry
+    $ResourceName = $MyInvocation.MyCommand.ModuleName -replace "MSFT_", ""
+    $CommandName = $MyInvocation.MyCommand
+    $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
+        -CommandName $CommandName `
+        -Parameters $PSBoundParameters
+    Add-M365DSCTelemetryEvent -Data $data
+    #endregion
+
+    Write-Verbose -Message "Testing configuration of EXOTransportConfig"
+
+    $CurrentValues = Get-TargetResource @PSBoundParameters
+
+    Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
+    Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
+
+    $ValuesToCheck = $PSBoundParameters
+    $ValuesToCheck.Remove('Credential') | Out-Null
+    $ValuesToCheck.Remove('ApplicationId') | Out-Null
+    $ValuesToCheck.Remove('TenantId') | Out-Null
+    $ValuesToCheck.Remove('CertificateThumbprint') | Out-Null
+    $ValuesToCheck.Remove('CertificatePath') | Out-Null
+    $ValuesToCheck.Remove('CertificatePassword') | Out-Null
+
+    if ($CurrentValues.Ensure -eq "Absent")
+    {
+        # In case transport config is missing at all for whatever reason.
+        $TestResult = $false
+    }
+    else
+    {
+        $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
+            -Source $($MyInvocation.MyCommand.Source) `
+            -DesiredValues $PSBoundParameters `
+            -ValuesToCheck $ValuesToCheck.Keys
+    }
+
+    Write-Verbose -Message "Test-TargetResource returned $TestResult"
+
+    return $TestResult
+}
+
+function Export-TargetResource
+{
+    [CmdletBinding()]
+    [OutputType([System.String])]
+    param
+    (
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $Credential,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword
+    )
+
+    $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
+        -InboundParameters $PSBoundParameters `
+        -SkipModuleReload $true
+
+    #Ensure the proper dependencies are installed in the current environment.
+    Confirm-M365DSCDependencies
+
+    #region Telemetry
+    $ResourceName = $MyInvocation.MyCommand.ModuleName -replace "MSFT_", ""
+    $CommandName = $MyInvocation.MyCommand
+    $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
+        -CommandName $CommandName `
+        -Parameters $PSBoundParameters
+    Add-M365DSCTelemetryEvent -Data $data
+    #endregion
+
+    try
+    {
+        $Params = @{
+            IsSingleInstance      = 'Yes'
+            Credential            = $Credential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
+            CertificatePassword   = $CertificatePassword
+            CertificatePath       = $CertificatePath
+        }
+
+        $Results = Get-TargetResource @Params
+        $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
+            -Results $Results
+        $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
+            -ConnectionMode $ConnectionMode `
+            -ModulePath $PSScriptRoot `
+            -Results $Results `
+            -Credential $Credential
+        $dscContent += $currentDSCBlock
+        Save-M365DSCPartialExport -Content $currentDSCBlock `
+            -FileName $Global:PartialExportFileName
+        Write-Host $Global:M365DSCEmojiGreenCheckMark
+        return $dscContent
+    }
+    catch
+    {
+        try
+        {
+            Write-Verbose -Message $_
+            $tenantIdValue = ""
+            if (-not [System.String]::IsNullOrEmpty($TenantId))
+            {
+                $tenantIdValue = $TenantId
+            }
+            elseif ($null -ne $Credential)
+            {
+                $tenantIdValue = $Credential.UserName.Split('@')[1]
+            }
+            Add-M365DSCEvent -Message $_ -EntryType 'Error' `
+                -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
+                -TenantId $tenantIdValue
+        }
+        catch
+        {
+            Write-Verbose -Message $_
+        }
+        return ""
+    }
+}
+
+Export-ModuleMember -Function *-TargetResource